Test methods and fix linters
