+++ enumerable_list.rb
@@ -1,0 +1,33 @@
+require_relative 'enumerable_module'
+class MyList
+  include MyEnumerable
+
+  def initialize(*enummembers)
+    @list = enummembers
+  end
+
+  def each(&enum_methods)
+    @list.each(&enum_methods)
+  end
+end
+
+# Check the solutions
+
+# Create our list
+list = MyList.new(1, 2, 3, 4)
+
+# Test #all?
+puts(list.all? { |e| e < 5 })
+# true
+puts(list.all? { |e| e > 5 })
+# false
+
+# Test #any?
+puts(list.any? { |e| e == 2 })
+# true
+puts(list.any? { |e| e == 5 })
+# false
+
+# Test #filter
+puts(list.filter(&:even?))
+# [2, 4]
