+++ enumerable_module.rb
